requirements.txt
certifi==2021.5.30
PyYAML==5.4.1
